Remove old normalize
