+++ client/payment-request/stripe-utils/index.js
@@ -1,3 +1,2 @@
-export * from './normalize';
 export * from './utils';
 export * from './load-stripe';
