--- client/payment-request/stripe-utils/normalize.js
@@ -1,218 +1,0 @@
-/**
- * @typedef {import('./type-defs').StripePaymentItem} StripePaymentItem
- * @typedef {import('./type-defs').StripeShippingOption} StripeShippingOption
- * @typedef {import('./type-defs').StripeShippingAddress} StripeShippingAddress
- * @typedef {import('./type-defs').StripePaymentResponse} StripePaymentResponse
- * @typedef {import('@woocommerce/type-defs/registered-payment-method-props').PreparedCartTotalItem} CartTotalItem
- * @typedef {import('@woocommerce/type-defs/cart').CartShippingOption} CartShippingOption
- * @typedef {import('@woocommerce/type-defs/shipping').ShippingAddress} CartShippingAddress
- * @typedef {import('@woocommerce/type-defs/billing').BillingData} CartBillingAddress
- */
-
-/**
- * Normalizes incoming cart total items for use as a displayItems with the
- * Stripe api.
- *
- * @param {CartTotalItem[]} cartTotalItems CartTotalItems to normalize
- * @param {boolean}         pending        Whether to mark items as pending or
- *                                         not
- *
- * @return {StripePaymentItem[]} An array of PaymentItems
- */
-const normalizeLineItems = ( cartTotalItems, pending = false ) => {
-	return cartTotalItems
-		.map( ( cartTotalItem ) => {
-			return cartTotalItem.value
-				? {
-						amount: cartTotalItem.value,
-						label: cartTotalItem.label,
-						pending,
-				  }
-				: false;
-		} )
-		.filter( Boolean );
-};
-
-/**
- * Normalizes incoming cart shipping option items for use as shipping options
- * with the Stripe api.
- *
- * @param {CartShippingOption[]}  shippingOptions An array of CartShippingOption items.
- *
- * @return {StripeShippingOption[]}  An array of Stripe shipping option items.
- */
-const normalizeShippingOptions = ( shippingOptions ) => {
-	const rates = shippingOptions[ 0 ].shipping_rates;
-	return rates.map( ( rate ) => {
-		return {
-			id: rate.rate_id,
-			label: rate.name,
-			detail: rate.description,
-			amount: parseInt( rate.price, 10 ),
-		};
-	} );
-};
-
-/**
- * Normalize shipping address information from stripe's address object to
- * the cart shipping address object shape.
- *
- * @param {StripeShippingAddress} shippingAddress Stripe's shipping address item
- *
- * @return {CartShippingAddress} The shipping address in the shape expected by
- * the cart.
- */
-const normalizeShippingAddressForCheckout = ( shippingAddress ) => {
-	return {
-		/* eslint-disable camelcase */
-		first_name:
-			shippingAddress.recipient
-				?.split( ' ' )
-				?.slice( 0, 1 )
-				?.join( ' ' ) ?? '',
-		last_name:
-			shippingAddress.recipient?.split( ' ' )?.slice( 1 )?.join( ' ' ) ??
-			'',
-		company: '',
-		address_1: shippingAddress.addressLine?.[ 0 ] ?? '',
-		address_2: shippingAddress.addressLine?.[ 1 ] ?? '',
-		city: shippingAddress.city ?? '',
-		state: shippingAddress.region ?? '',
-		country: shippingAddress.country ?? '',
-		postcode: shippingAddress.postalCode?.replace( ' ', '' ) ?? '',
-		/* eslint-enable camelcase */
-	};
-};
-
-const normalizeOrderDataForCheckout = ( paymentData ) => {
-	const name = paymentData.paymentMethod.billing_details.name;
-	const email = paymentData.paymentMethod.billing_details.email;
-	const phone = paymentData.paymentMethod.billing_details.phone;
-	const billing = paymentData.paymentMethod.billing_details.address;
-	const shipping = paymentData.shippingAddress ?? {};
-
-	return {
-		/* eslint-disable camelcase */
-		billing_first_name:
-			name?.split( ' ' )?.slice( 0, 1 )?.join( ' ' ) ?? '',
-		billing_last_name: name?.split( ' ' )?.slice( 1 )?.join( ' ' ) ?? '',
-		billing_company: billing.organization,
-		billing_email: email ?? paymentData.payerEmail,
-		billing_phone:
-			phone ?? paymentData.payerPhone.replace( '/[() -]/g', '' ),
-		billing_country: billing.country ?? '',
-		billing_address_1: billing.line1 ?? '',
-		billing_address_2: billing.line2 ?? '',
-		billing_city: billing.city ?? '',
-		billing_state: billing.state ?? '',
-		billing_postcode: billing.postal_code ?? '',
-		shipping_first_name:
-			shipping?.recipient?.split( ' ' )?.slice( 0, 1 )?.join( ' ' ) ?? '',
-		shipping_last_name:
-			shipping?.recipient?.split( ' ' )?.slice( 1 )?.join( ' ' ) ?? '',
-		// shipping_company: shipping.organization ?? '',
-		shipping_country: shipping.country ?? '',
-		shipping_address_1: shipping.addressLine?.[ 0 ] ?? '',
-		shipping_address_2: shipping.addressLine?.[ 1 ] ?? '',
-		shipping_city: shipping.city ?? '',
-		shipping_state: shipping.region ?? '',
-		shipping_postcode: shipping.postalCode ?? '',
-		shipping_method: [
-			null === paymentData.shippingOption
-				? null
-				: paymentData.shippingOption.id,
-		],
-		order_comments: '',
-		payment_method: 'woocommerce_payments', // - TODO: Replace by constant
-		ship_to_different_address: 1,
-		terms: 1,
-		'wcpay-payment-method': paymentData.paymentMethod.id,
-		/* eslint-enable camelcase */
-	};
-};
-
-/**
- * Normalizes shipping option shape selection from Stripe's shipping option
- * object to the expected shape for cart shipping option selections.
- *
- * @param {StripeShippingOption} shippingOption The customer's selected shipping
- *                                              option.
- *
- * @return {string[]}  An array of ids (in this case will just be one)
- */
-const normalizeShippingOptionSelectionsForCheckout = ( shippingOption ) => {
-	return shippingOption.id;
-};
-
-/**
- * Returns the billing data extracted from the stripe payment response to the
- * CartBillingData shape.
- *
- * @param {StripePaymentResponse} paymentResponse Stripe's payment response
- *                                                object.
- *
- * @return {CartBillingAddress} The cart billing data
- */
-const getBillingData = ( paymentResponse ) => {
-	const source = paymentResponse.source;
-	const name = source?.owner?.name;
-	const billing = source?.owner?.address;
-	const payerEmail = paymentResponse.payerEmail ?? '';
-	const payerPhone = paymentResponse.payerPhone ?? '';
-	return {
-		/* eslint-disable camelcase */
-		first_name: name?.split( ' ' )?.slice( 0, 1 )?.join( ' ' ) ?? '',
-		last_name: name?.split( ' ' )?.slice( 1 )?.join( ' ' ) ?? '',
-		email: source?.owner?.email ?? payerEmail,
-		phone: source?.owner?.phone ?? payerPhone.replace( '/[() -]/g', '' ),
-		country: billing?.country ?? '',
-		address_1: billing?.line1 ?? '',
-		address_2: billing?.line2 ?? '',
-		city: billing?.city ?? '',
-		state: billing?.state ?? '',
-		postcode: billing?.postal_code ?? '',
-		company: '',
-		/* eslint-enable camelcase */
-	};
-};
-
-/**
- * This returns extra payment method data to add to the payment method update
- * request made by the checkout processor.
- *
- * @param {StripePaymentResponse} paymentResponse    A stripe payment response
- *                                                   object.
- * @param {string}                paymentRequestType The payment request type
- *                                                   used for payment.
- *
- * @return {Object} An object with the extra payment data.
- */
-const getPaymentMethodData = ( paymentResponse, paymentRequestType ) => {
-	return {
-		/* eslint-disable camelcase */
-		payment_method: 'woocommerce_payments',
-		stripe_source: paymentResponse.source?.id,
-		/* eslint-enable camelcase */
-	};
-};
-
-const getShippingData = ( paymentResponse ) => {
-	return paymentResponse.shippingAddress
-		? {
-				address: normalizeShippingAddressForCheckout(
-					paymentResponse.shippingAddress
-				),
-		  }
-		: null;
-};
-
-export {
-	normalizeLineItems,
-	normalizeShippingOptions,
-	normalizeShippingAddressForCheckout,
-	normalizeShippingOptionSelectionsForCheckout,
-	normalizeOrderDataForCheckout,
-	getBillingData,
-	getPaymentMethodData,
-	getShippingData,
-};
